requirements.txt
cryptography
sqlcipher3-binary